fix calculateRemainingAndSignedMembers logic
